Relocate GSP observation kit to grails-gsp-core; instrument compilation (gsp.compile)
Moves the observation kit (Context/Convention/Documentation) from grails-web-gsp down
to grails-gsp-core so the template engine — which lives in core, below the web layer,
and which core cannot import upward from web-gsp — can use it too. View/template/layout
importers are repointed to the new org.grails.gsp.observation package.

Instruments GroovyPagesTemplateEngine.buildPageMetaInfo with a 'gsp.compile' observation
(observeChecked). Compilation only happens on a template cache miss, so the observation's
count is the compile/miss rate and its timer is the compile latency; tagged with the GSP
page name. The ObservationRegistry is resolved from the application context (NOOP fallback).
